Create ItemEvent and move some events there
