+++ src/main/java/nova/core/item/ItemManager.java
@@ -22,8 +22,7 @@ package nova.core.item;
 
 import nova.core.block.BlockFactory;
 import nova.core.block.BlockManager;
-import nova.core.event.bus.CancelableEvent;
-import nova.core.item.event.ItemIDNotFoundEvent;
+import nova.core.event.ItemEvent;
 import nova.core.util.registry.FactoryManager;
 import nova.core.util.registry.Registry;
 import nova.internal.core.Game;
@@ -52,9 +51,10 @@ public class ItemManager extends FactoryManager<ItemManager, Item, ItemFactory> 
 
 	@Override
 	public ItemFactory register(ItemFactory factory) {
-		registry.register(factory);
-		Game.events().publish(new ItemRegistrationEvent(factory));
-		return factory;
+		ItemEvent.Register event = new ItemEvent.Register(factory);
+		Game.events().publish(event);
+		registry.register(event.itemFactory);
+		return event.itemFactory;
 	}
 
 	public ItemFactory getItemFromBlock(BlockFactory block) {
@@ -68,7 +68,7 @@ public class ItemManager extends FactoryManager<ItemManager, Item, ItemFactory> 
 	@Override
 	public Optional<ItemFactory> get(String name) {
 		if (!registry.contains(name)) {
-			ItemIDNotFoundEvent event = new ItemIDNotFoundEvent(name);
+			ItemEvent.IDNotFound event = new ItemEvent.IDNotFound(name);
 			Game.events().publish(event);
 
 			if (event.getRemappedFactory() != null) {
@@ -77,15 +77,6 @@ public class ItemManager extends FactoryManager<ItemManager, Item, ItemFactory> 
 		}
 
 		return registry.get(name);
-	}
-
-	//TODO: Move to item event
-	public class ItemRegistrationEvent extends CancelableEvent {
-		public final ItemFactory itemFactory;
-
-		public ItemRegistrationEvent(ItemFactory itemFactory) {
-			this.itemFactory = itemFactory;
-		}
 	}
 
 	@Override
